requirements.txt
<<<<<<< HEAD
fastapi
uvicorn[standard]
SQLAlchemy
python-jose
bcrypt
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
email-validator
celery
redis
Pillow
opencv-python
transformers
sentence-transformers
openai
requests
pytest
python-dotenv
loguru
httpx
psycopg2-binary
alembic
=======
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
celery==5.4.0
redis==5.2.0
pillow==11.0.0
opencv-python==4.10.0.84
transformers==4.46.2
sentence-transformers==3.2.1
openai==1.54.3
requests==2.32.3
pytest==8.3.3
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.2
psycopg2-binary==2.9.10
pgvector==0.3.5
alembic==1.13.3
numpy==1.26.4
torch==2.5.1
>>>>>>> f3fded8fa24db6078ca24e84e18bb276abe04a3a
